+++ RELEASENOTES.md
@@ -2,6 +2,18 @@
 
 Releases, starting with 9/2/2021, are listed with the most recent release at the top.
 
+## NuGet Version 0.101.3
+
+__API Changes__:
+
+Introduced `InferenceMode`, a block-based scoping class for optimizing TorchSharp model inference by disabling gradient computation and enhancing performance.<br/>
+Added `Tensor.to_type()` conversion aliases for short, half, bfloat16, cfloat, and cdouble.<br/>
+Added `Module.to()` conversion aliases for all the scalar types.<br/>
+
+__Bug Fixes__:
+
+#1154 : `mu_product` was not initialized in `NAdam` optimizer
+
 ## NuGet Version 0.101.2
 
 __API Changes__: